+++ protobuf/lib/src/protobuf/field_info.dart
@@ -8,6 +8,9 @@ part of protobuf;
 class FieldInfo<T> {
   FrozenPbList<T> _emptyList;
 
+  /// Name of this field as the `json_name` reported by protoc.
+  ///
+  /// This will typically be in camel case.
   final String name;
 
   /// The name of this field as written in the proto-definition.
@@ -188,15 +191,25 @@ String _unCamelCase(String name) {
 }
 
 class MapFieldInfo<K, V> extends FieldInfo<PbMap<K, V>> {
-  int keyFieldType;
-  int valueFieldType;
-  CreateBuilderFunc valueCreator;
+  final int keyFieldType;
+  final int valueFieldType;
 
-  // BuilderInfo used when creating a field set for a map field.
+  /// Creates a new empty instance of the value type.
+  ///
+  /// `null` if the value type is not a Message type.
+  final CreateBuilderFunc valueCreator;
+
   final BuilderInfo _mapEntryBuilderInfo;
 
-  MapFieldInfo(String name, int tagNumber, int index, int type,
-      this.keyFieldType, this.valueFieldType, this._mapEntryBuilderInfo,
+  MapFieldInfo(
+      String name,
+      int tagNumber,
+      int index,
+      int type,
+      this.keyFieldType,
+      this.valueFieldType,
+      this._mapEntryBuilderInfo,
+      this.valueCreator,
       {String protoName})
       : super(name, tagNumber, index, type,
             defaultOrMaker: () =>
@@ -207,6 +220,9 @@ class MapFieldInfo<K, V> extends FieldInfo<PbMap<K, V>> {
     assert(_isMapField(type));
     assert(!_isEnum(type) || valueOf != null);
   }
+
+  FieldInfo get valueFieldInfo =>
+      _mapEntryBuilderInfo.fieldInfo[PbMap._valueFieldNumber];
 
   Map<K, V> _ensureMapField(_FieldSet fs) {
     return fs._ensureMapField<K, V>(this);