+++ protoc_plugin/lib/message_generator.dart
@@ -291,6 +291,10 @@ class MessageGenerator extends ProtobufContainer {
     String packageClause = package == ''
         ? ''
         : ', package: const $_protobufImportPrefix.PackageName(\'$package\')';
+    String proto3JsonClause = (mixin?.hasProto3JsonHelpers ?? false)
+        ? ', toProto3Json: $_mixinImportPrefix.${mixin.name}.toProto3JsonHelper, '
+            'fromProto3Json: $_mixinImportPrefix.${mixin.name}.fromProto3JsonHelper'
+        : '';
     out.addAnnotatedBlock(
         'class ${classname} extends $_protobufImportPrefix.GeneratedMessage${mixinClause} {',
         '}', [
@@ -312,7 +316,9 @@ class MessageGenerator extends ProtobufContainer {
       }
       out.addBlock(
           'static final $_protobufImportPrefix.BuilderInfo _i = '
-              '$_protobufImportPrefix.BuilderInfo(\'${messageName}\'$packageClause)',
+              '$_protobufImportPrefix.BuilderInfo(\'${messageName}\'$packageClause'
+              ', createEmptyInstance: create'
+              '$proto3JsonClause)',
           ';', () {
         for (int oneof = 0; oneof < _oneofFields.length; oneof++) {
           List<int> tags =